requirements.txt
certifi==2020.12.5
click==7.1.2
cloudinary==1.25.0
fastapi==0.65.1
h11==0.12.0
numpy==1.20.3
opencv-python-headless==4.5.2.52
Pillow==8.2.0
pydantic==1.8.2
pymongo==3.11.4
python-dotenv==0.17.1
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
tflite-runtime==2.5.0
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
-i https://google-coral.github.io/py-repo/ tflite_runtime